refactor: add comments and rename dir
