+++ internal/engine/worker/factory.go
@@ -6,6 +6,7 @@ import (
 	"github.com/rss3-network/node/config"
 	"github.com/rss3-network/node/internal/database"
 	"github.com/rss3-network/node/internal/engine"
+	"github.com/rss3-network/node/internal/engine/worker/contract/1inch"
 	"github.com/rss3-network/node/internal/engine/worker/contract/aave"
 	"github.com/rss3-network/node/internal/engine/worker/contract/aavegotchi"
 	"github.com/rss3-network/node/internal/engine/worker/contract/crossbell"
@@ -18,7 +19,6 @@ import (
 	"github.com/rss3-network/node/internal/engine/worker/contract/matters"
 	"github.com/rss3-network/node/internal/engine/worker/contract/mirror"
 	"github.com/rss3-network/node/internal/engine/worker/contract/momoka"
-	"github.com/rss3-network/node/internal/engine/worker/contract/oneinch"
 	"github.com/rss3-network/node/internal/engine/worker/contract/opensea"
 	"github.com/rss3-network/node/internal/engine/worker/contract/optimism"
 	"github.com/rss3-network/node/internal/engine/worker/contract/paragraph"
